ref: split tool modules into tools/ package
