+++ langchain_apify/__init__.py
@@ -3,9 +3,16 @@ from __future__ import annotations
 from importlib import metadata
 from typing import TYPE_CHECKING
 
-from langchain_apify._actor_tools import (
+from langchain_apify.document_loaders import ApifyCrawlLoader, ApifyDatasetLoader
+from langchain_apify.retrievers import ApifySearchRetriever
+from langchain_apify.tools import (
+    APIFY_CORE_TOOLS,
+    APIFY_SEARCH_TOOLS,
+    APIFY_SOCIAL_TOOLS,
+    ApifyActorsTool,
     ApifyEcommerceScraperTool,
     ApifyFacebookPostsScraperTool,
+    ApifyGetDatasetItemsTool,
     ApifyGoogleMapsTool,
     ApifyGoogleSearchTool,
     ApifyInstagramScraperTool,
@@ -13,21 +20,15 @@ from langchain_apify._actor_tools import (
     ApifyLinkedInProfilePostsTool,
     ApifyLinkedInProfileSearchTool,
     ApifyRAGWebBrowserTool,
-    ApifyTikTokScraperTool,
-    ApifyTwitterScraperTool,
-    ApifyWebCrawlerTool,
-    ApifyYouTubeScraperTool,
-)
-from langchain_apify.document_loaders import ApifyCrawlLoader, ApifyDatasetLoader
-from langchain_apify.retrievers import ApifySearchRetriever
-from langchain_apify.tools import (
-    ApifyActorsTool,
-    ApifyGetDatasetItemsTool,
     ApifyRunActorAndGetDatasetTool,
     ApifyRunActorTool,
     ApifyRunTaskAndGetDatasetTool,
     ApifyRunTaskTool,
     ApifyScrapeUrlTool,
+    ApifyTikTokScraperTool,
+    ApifyTwitterScraperTool,
+    ApifyWebCrawlerTool,
+    ApifyYouTubeScraperTool,
 )
 from langchain_apify.wrappers import ApifyWrapper
 
@@ -35,42 +36,10 @@ if TYPE_CHECKING:
     from langchain_core.tools import BaseTool
 
 try:
-    __version__ = metadata.version(__package__)
+    __version__ = metadata.version(__package__) if __package__ is not None else ''
 except metadata.PackageNotFoundError:
     __version__ = ''
 del metadata  # optional, avoids polluting the results of dir(__package__)
-
-# Convenience tool-class lists for selective agent binding.
-# Binding all tools at once overwhelms the LLM context window;
-# pick the group(s) relevant to your use case.
-
-APIFY_CORE_TOOLS: list[type[BaseTool]] = [
-    ApifyRunActorTool,
-    ApifyGetDatasetItemsTool,
-    ApifyRunActorAndGetDatasetTool,
-    ApifyScrapeUrlTool,
-    ApifyRunTaskTool,
-    ApifyRunTaskAndGetDatasetTool,
-]
-
-APIFY_SOCIAL_TOOLS: list[type[BaseTool]] = [
-    ApifyInstagramScraperTool,
-    ApifyLinkedInProfilePostsTool,
-    ApifyLinkedInProfileSearchTool,
-    ApifyLinkedInProfileDetailTool,
-    ApifyTwitterScraperTool,
-    ApifyTikTokScraperTool,
-    ApifyFacebookPostsScraperTool,
-]
-
-APIFY_SEARCH_TOOLS: list[type[BaseTool]] = [
-    ApifyGoogleSearchTool,
-    ApifyWebCrawlerTool,
-    ApifyRAGWebBrowserTool,
-    ApifyGoogleMapsTool,
-    ApifyYouTubeScraperTool,
-    ApifyEcommerceScraperTool,
-]
 
 __all__ = [
     # Existing components (backward-compatible)
